1.8.5-2

+++ libs/linglong/src/linglong/package_manager/package_manager.cpp
@@ -19,6 +19,7 @@
 #include "linglong/utils/command/env.h"
 #include "linglong/utils/configure.h"
 #include "linglong/utils/finally/finally.h"
+#include "linglong/utils/hooks.h"
 #include "linglong/utils/packageinfo_handler.h"
 #include "linglong/utils/serialize/json.h"
 #include "linglong/utils/transaction.h"
@@ -497,8 +498,7 @@ void PackageManager::setConfiguration(const QVariantMap &parameters) noexcept
 QVariantMap PackageManager::installFromLayer(const QDBusUnixFileDescriptor &fd,
                                              const api::types::v1::CommonOptions &options) noexcept
 {
-    auto layerFileRet =
-      package::LayerFile::New(QString("/proc/%1/fd/%2").arg(getpid()).arg(fd.fileDescriptor()));
+    auto layerFileRet = package::LayerFile::New(fd.fileDescriptor());
     if (!layerFileRet) {
         return toDBusReply(layerFileRet);
     }
@@ -726,6 +726,12 @@ QVariantMap PackageManager::installFromLayer(const QDBusUnixFileDescriptor &fd,
                           << "after install" << packageRef.toString() << ":"
                           << ret.error().message();
           }
+
+          ret = executePostInstallHooks(*newRef);
+          if (!ret) {
+              taskRef.reportError(std::move(ret).error());
+              return;
+          }
       };
 
     auto refSpec =
@@ -752,8 +758,23 @@ QVariantMap PackageManager::installFromLayer(const QDBusUnixFileDescriptor &fd,
 QVariantMap PackageManager::installFromUAB(const QDBusUnixFileDescriptor &fd,
                                            const api::types::v1::CommonOptions &options) noexcept
 {
-    auto uabRet = package::UABFile::loadFromFile(
-      QString("/proc/%1/fd/%2").arg(getpid()).arg(fd.fileDescriptor()));
+    if (!fd.isValid()) {
+        return toDBusReply(-1, "invalid file descriptor");
+    }
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return toDBusReply(ret);
+    }
+
+    ret = installHookManager->executeInstallHooks(fd.fileDescriptor());
+    if (!ret) {
+        return toDBusReply(-1, "uab package signature verification failed.");
+    }
+
+    auto uabRet = package::UABFile::loadFromFile(fd.fileDescriptor());
     if (!uabRet) {
         return toDBusReply(uabRet);
     }
@@ -1055,6 +1076,12 @@ QVariantMap PackageManager::installFromUAB(const QDBusUnixFileDescriptor &fd,
                                     "Failed to generate some cache.\n" + result.error().message());
                 return;
             }
+        }
+
+        auto ret = executePostInstallHooks(newAppRef);
+        if (!ret) {
+            taskRef.reportError(std::move(ret).error());
+            return;
         }
 
         transaction.commit();
@@ -1365,6 +1392,13 @@ void PackageManager::Install(PackageTask &taskContext,
         }
     }
 
+    auto ret = executePostInstallHooks(newRef);
+    if (!ret) {
+        taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                "Failed to execute postInstall hooks.\n" + ret.error().message());
+        return;
+    }
+
     transaction.commit();
     taskContext.updateState(linglong::api::types::v1::State::Succeed,
                             "Install " + newRef.toString() + " success");
@@ -1633,6 +1667,11 @@ void PackageManager::Uninstall(PackageTask &taskContext,
         return;
     }
 
+    auto ret = executePostUninstallHooks(ref);
+    if (!ret) {
+        qWarning() << "failed to execute postUninstall hooks" << ret.error();
+    }
+
     transaction.commit();
 
     taskContext.updateState(linglong::api::types::v1::State::Succeed,
@@ -1731,7 +1770,7 @@ auto PackageManager::Update(const QVariantMap &parameters) noexcept -> QVariantM
       refSpecs,
       [this, upgrades = std::move(upgrades)](PackageTask &taskRef) {
           for (const auto &[reference, newReference] : upgrades) {
-              if (isTaskDone(taskRef.subState())) {
+              if (taskRef.subState() == linglong::api::types::v1::SubState::AllDone) {
                   return;
               }
 
@@ -1804,6 +1843,12 @@ void PackageManager::Update(PackageTask &taskContext,
         if (!ret) {
             qCritical() << "remove after install of ref" << ref.toString()
                         << "failed:" << ret.error().message();
+            return;
+        }
+
+        ret = executePostInstallHooks(newRef);
+        if (!ret) {
+            qCritical() << "failed to execute post install hooks" << ret.error().message();
             return;
         }
 
@@ -1906,8 +1951,21 @@ void PackageManager::pullDependency(PackageTask &taskContext,
                 return;
             }
 
+            auto ret = executePostInstallHooks(*runtime);
+            if (!ret) {
+                taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                        LINGLONG_ERRV(ret).message());
+                return;
+            }
+
             transaction.addRollBack([this, runtimeRef = *runtime, module]() noexcept {
                 auto result = this->repo.remove(runtimeRef, module);
+                if (!result) {
+                    qCritical() << result.error();
+                    Q_ASSERT(false);
+                }
+
+                result = executePostUninstallHooks(runtimeRef);
                 if (!result) {
                     qCritical() << result.error();
                     Q_ASSERT(false);
@@ -1947,6 +2005,27 @@ void PackageManager::pullDependency(PackageTask &taskContext,
         if (isTaskDone(taskContext.subState())) {
             return;
         }
+
+        auto ret = executePostInstallHooks(*base);
+        if (!ret) {
+            taskContext.updateState(linglong::api::types::v1::State::Failed,
+                                    LINGLONG_ERRV(ret).message());
+            return;
+        }
+
+        transaction.addRollBack([this, baseRef = *base, module]() noexcept {
+            auto result = this->repo.remove(baseRef, module);
+            if (!result) {
+                qCritical() << result.error();
+                Q_ASSERT(false);
+            }
+
+            result = executePostUninstallHooks(baseRef);
+            if (!result) {
+                qCritical() << result.error();
+                Q_ASSERT(false);
+            }
+        });
     }
 
     transaction.commit();
@@ -2405,6 +2484,53 @@ auto PackageManager::GenerateCache(const QString &reference) noexcept -> QVarian
       .message = "",
     });
     return result;
+}
+
+utils::error::Result<void>
+PackageManager::executePostInstallHooks(const package::Reference &ref) noexcept
+{
+    LINGLONG_TRACE("execute post install hooks for: " + ref.toString());
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    auto layerDir = this->repo.getLayerDir(ref);
+    if (!layerDir) {
+        return LINGLONG_ERR(layerDir);
+    }
+
+    auto appPath = layerDir->absolutePath();
+
+    ret = installHookManager->executePostInstallHooks(ref.id.toStdString(), appPath.toStdString());
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    return LINGLONG_OK;
+}
+
+utils::error::Result<void>
+PackageManager::executePostUninstallHooks(const package::Reference &ref) noexcept
+{
+    LINGLONG_TRACE("execute post uninstall hooks for: " + ref.toString());
+
+    std::unique_ptr<utils::InstallHookManager> installHookManager =
+      std::make_unique<utils::InstallHookManager>();
+    auto ret = installHookManager->parseInstallHooks();
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    ret = installHookManager->executePostUninstallHooks(ref.id.toStdString());
+    if (!ret) {
+        return LINGLONG_ERR(ret);
+    }
+
+    return LINGLONG_OK;
 }
 
 } // namespace linglong::service
